# docs/swagger.yaml
basePath: /v1/api
definitions:
...
  go-app_internal_dto.ConnectionDto:
    properties:
      receive_id:
        type: string
      request_id:
        type: string
    type: object
...
  go-app_internal_dto.CreateGroupDto:
    properties:
      group_name:
        type: string
      member_count:
        description: MemberIds   []string `json:"member_ids" binding:"required"`
        type: integer
      owner_id:
        type: string
    required:
    - group_name
    - member_count
    - owner_id
    type: object
...
  go-app_internal_dto.Participants:
    properties:
      participants:
        items:
          type: string
        type: array
    type: object
...
  go-app_internal_dto.UpdateGroupDto:
    properties:
      group_name:
        type: string
      member_count:
        type: integer
      status:
        $ref: '#/definitions/go-app_internal_schema.GroupStatus'
    type: object
...
  go-app_internal_schema.GroupStatus:
    enum:
    - ACTIVE
    - DELETED
    type: string
    x-enum-varnames:
    - GroupStatusActive
    - GroupStatusDeleted
host: localhost:8081
info:
  contact: {}
  description: This is a sample server API for Go App.
  title: Go App API
  version: "1.0"
paths:
...
  /connection:
    post:
      consumes:
      - application/json
      description: Create a new connection/friend request
      parameters:
      - description: Connection Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.ConnectionDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create connection
      tags:
      - connection
  /connection/{connection-id}:
    delete:
      description: Delete a connection by ID
      parameters:
      - description: Connection ID
        in: path
        name: connection-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete connection
      tags:
      - connection
  /connection/{connection-id}/respond:
    put:
      description: Accept or reject a connection request
      parameters:
      - description: Connection ID
        in: path
        name: connection-id
        required: true
        type: string
      - description: Status (accepted, rejected)
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Respond to connection request
      tags:
      - connection
  /connection/detail:
    post:
      consumes:
      - application/json
      description: Get connection details by participants
      parameters:
      - description: Participants Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.Participants'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get connection details
      tags:
      - connection
  /connection/user:
    get:
      description: Get connections for the current user, optionally filtered by status
      parameters:
      - description: Connection Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get user connections
      tags:
      - connection
  /group:
    post:
      consumes:
      - application/json
      description: Create a new group along with channel, members, and unreads
      parameters:
      - description: Create Group Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.CreateGroupDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create a new group
      tags:
      - group
  /group/{group-id}:
    delete:
      description: Delete a group by its ID
      parameters:
      - description: Group ID
        in: path
        name: group-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete a group
      tags:
      - group
    put:
      consumes:
      - application/json
      description: Update information of an existing group
      parameters:
      - description: Group ID
        in: path
        name: group-id
        required: true
        type: string
      - description: Update Group Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.UpdateGroupDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update group info
      tags:
      - group
...
swagger: "2.0"